Adjust Energi Fyn - SpotEl
According to https://www.energifyn.dk/privat/el/spotel/ the addition to the spot price is now 4,50 øre/kWh and not 4,00 øre/kWh. Adjust this to reflect the real price. There is also a subscription fee of 15 kr / month. Add that to fees and remove the subscription fee from the kWh price as it seems not to present anymore.