Adjust Energi Fyn - SpotEl (#83)
* Adjust Energi Fyn - SpotEl

According to https://www.energifyn.dk/privat/el/spotel/ the addition to the spot price is now 4,50 øre/kWh and not 4,00 øre/kWh. Adjust this to reflect the real price. There is also a subscription fee of 15 kr / month. Add that to fees and remove the subscription fee from the kWh price as it seems not to present anymore.

* Readd subscribtion

Readded the subscription as it was the subscription fee to Vores Elnet.  Prices can be found at https://www.voreselnet.dk/priser, which is currently 324 kr / year.

The right fix for this is to add it to the `tariffs` list.

* Only add subscription

* Adjust price to be without VAT